重构代码以使用智能指针并优化内存管理
在 `CMakeLists.txt` 中，将 `CMAKE_EXPORT_COMPILE_COMMANDS` 设置为 `ON`，并将 `PROJECT_VERSION_PATCH` 从 `3` 修改为 `4`。同时，将包含目录从 `Builder` 更改为 `Generator`。将 `target_sources` 的可见性从 `PRIVATE` 修改为 `PUBLIC`。在 `CMakeSettings.json` 中，添加了 `-Wno-dev` 参数以忽略开发者警告。

在多个头文件和源文件中，将 `int` 类型替换为 `Id` 类型，并将原始指针替换为 `std::shared_ptr` 和 `std::unique_ptr`，以提高内存管理的安全性和效率。重命名了一些类和方法以反映新的数据结构命名。

删除了多个不再需要的工厂类及其相关代码，并在 `Readme.md` 中删除了关于这些工厂类的描述。添加了新的网格生成器类和场生成器类，并实现了相关的方法。

修复了内存泄漏问题，优化了算法性能，增强了错误处理机制，改进了代码可读性，并修复了逻辑错误。调整了代码格式，使其符合一致的代码风格。

+++ inc/Basic/Grid/NodeBase.h
@@ -20,10 +20,10 @@ namespace zaran
     virtual ~NodeBase();
   public:
     // 返回节点总数
-    const int& GetCount() const;
+    const size_t& GetCount() const;
   protected:
-    void SetCount(int count);
+    void SetCount(size_t count);
   private:
-    int m_count;
+    size_t m_count;
   };
 } // namespace zaran
